Adds ability to listen for clicks on overlays
This PR adds the ability to pass a `click` option into the `drawOverlay`
method in order to perform some action when an overlay's DOM element is clicked.
An accompanying test has been added.

When adding complex markers to a map, ones that use custom HTML/CSS, for
instance, it is necessary to use overlays rather than markers. That being the
case, it is desirable that a developer be able to easily listen to click events
on those overlays the same as they would on a standard marker.