Switch to relative links for internal pages
This switches all of the links in the logo and in a few of the root
pages to use relative links instead of hard-coding the paths.  The
hard-coded paths made it incredibly difficult to view the
documentation locally, as your are constantly switched to the live
version of the documentation.