Fixed mixed content warnings under TLS
This fixed the mixed content warning in the map types page that
resulted from using the non-SSL endpoints for the OpenStreetMap
Tile API. The image paths have been switched, so there should no
longer be a mixed content warning.

This also removes the CloudMade example map, as the API key does
not appear to be active so it has not been loading the map at all.